+++ client/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "numbernumbergame",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "async": "^3.1.0",
+    "babel": "^6.23.0",
+    "bootstrap": "^4.3.1",
+    "connect": "^3.7.0",
+    "hooks": "^0.3.2",
+    "react": "^16.11.0",
+    "react-bootstrap": "^1.0.0-beta.14",
+    "react-dom": "^16.11.0",
+    "react-redux": "^7.1.1",
+    "react-scripts": "0.9.5",
+    "redux": "^4.0.4"
+  },
+  "devDependencies": {},
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
+  }
+}
