+++ client/src/components/player/index.js
