+++ client/src/components/player/reducer.js
