+++ client/src/components/styles/StyledCell.js
