+++ client/src/components/styles/StyledDisplay.js
