+++ client/src/components/styles/StyledGameOver.js
