+++ client/src/components/styles/StyledStage.js
