+++ package.json
@@ -1,25 +1,31 @@
 {
   "name": "numbernumbergame",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "async": "^3.1.0",
-    "babel": "^6.23.0",
-    "bootstrap": "^4.3.1",
-    "connect": "^3.7.0",
-    "hooks": "^0.3.2",
-    "react": "^16.11.0",
-    "react-bootstrap": "^1.0.0-beta.14",
-    "react-dom": "^16.11.0",
-    "react-redux": "^7.1.1",
-    "react-scripts": "0.9.5",
-    "redux": "^4.0.4"
-  },
-  "devDependencies": {},
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  }
+    "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
+    "start:prod": "node server.js",
+    "start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"npm run client\"",
+    "client": "cd client && npm run start",
+    "seed": "node scripts/seedDB.js",
+    "install": "cd client && npm install",
+    "build": "cd client && npm run build"
+  },
+  "dependencies": {
+    "concurrently": "^5.0.0",
+    "express": "^4.17.1",
+    "if-env": "^1.0.4",
+    "socket.io": "^2.3.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/lobo89x/numbernumbergame.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/lobo89x/numbernumbergame/issues"
+  },
+  "homepage": "https://github.com/lobo89x/numbernumbergame#readme"
 }
